+++ packages/sources/src/components/ozon-content/ozon-content.args.ts
@@ -5,7 +5,8 @@ import { OzonContent } from "./ozon-content.models";
 
 export interface OzonContentArgs {
   content: string;
-  inline?: boolean;
+  inline: boolean;
+  deleted: boolean;
   onAnchorClick: HandlerFunction;
 }
 
@@ -20,6 +21,11 @@ export const ozonContentArgTypes: ArgTypes<OzonContentArgs> = {
       type: 'boolean'
     }
   },
+  deleted: {
+    control: {
+      type: 'boolean'
+    }
+  },
   onAnchorClick: {
     ...noControl,
     action: "anchorClick",
@@ -30,6 +36,7 @@ export function ozonContentArgsMapper(a: OzonContentArgs): OzonContent {
   return {
     content: a.content,
     inline: a.inline,
+    deleted: a.deleted,
     onAnchorClick: (e: any) => a.onAnchorClick(e.detail),
   };
 }