+++ packages/sources/src/components/ozon-content/ozon-content.models.ts
@@ -1,5 +1,6 @@
 export interface OzonContent {
   content: string;
   inline?: boolean;
+  deleted?: boolean;
   onAnchorClick: (e: unknown) => void
 }