+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * **core:** `<dso-ozon-content inline>` en ondersteuning voor `<Opschrift>` ([#1530](https://github.com/dso-toolkit/dso-toolkit/issues/1530))
+* **core:** `<dso-ozon-content deleted>` voor het markeren van verwijderde content ([#1518](https://github.com/dso-toolkit/dso-toolkit/issues/1518))
 * **sources + dso-toolkit + css + core:** Badge; `badge-outline` toevoegen ([#1534]https://github.com/dso-toolkit/dso-toolkit/issues/1534)
 
 ### Removed